miniflux/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "miniflux>=1.1.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
